+++ framework/Widget.h
@@ -20,7 +20,7 @@ public:
 
 	void scale(sf::Vector2f scale_proportions);
 
-	virtual bool containsPoint(sf::Vector2f point_coordinates)const = 0;
+	virtual bool containsPoint(sf::Vector2i point_coordinates)const = 0;
 private:
 	const std::string name_;
 	sf::Vector2i position_;
